Refactor: 전체적인 폴더구조 변경
