+++ src/components/post/PostList.tsx
@@ -4,12 +4,13 @@ import { useNavigate } from 'react-router-dom';
 import { Button } from '@components/common/Button';
 import DataStateHandler from '@components/common/DataStateHandler';
 import Pagination from '@components/common/Pagination';
-import PostListSkeleton from '@components/skeleton/PostListSkeleton';
+import { CommentIcon } from '@components/icons';
+import PostListSkeleton from '@components/skeleton/post/PostListSkeleton';
 
 import { API_DEFAULTS } from '@constants/api';
 import { PATH } from '@constants/path';
 
-import { usePostsQuery } from '@hooks/queries/usePostsQuery';
+import { usePostsQuery } from '@hooks/queries/post/usePostsQuery';
 
 import { useAuthStore } from '@stores/useAuthStore';
 
@@ -132,31 +133,7 @@ const PostList = () => {
                           className="flex items-center gap-1"
                           aria-label={`댓글 ${post.commentCount}개`}
                         >
-                          <svg
-                            width="23"
-                            height="19"
-                            viewBox="0 0 23 19"
-                            fill="none"
-                            xmlns="http://www.w3.org/2000/svg"
-                            aria-label="댓글 아이콘"
-                          >
-                            <g clipPath="url(#clip0_1_6044)">
-                              <path
-                                d="M7.3071 18.5C7.72951 18.5 8.02262 18.287 8.54848 17.8446L11.5226 15.3293H17.0571C19.6261 15.3293 21.0054 13.9774 21.0054 11.577V5.3011C21.0054 2.90055 19.6261 1.54871 17.0571 1.54871H5.95365C3.38468 1.54871 2.00537 2.89236 2.00537 5.3011V11.577C2.00537 13.9857 3.38468 15.3293 5.95365 15.3293H6.36745V17.4595C6.36745 18.0904 6.70365 18.5 7.3071 18.5ZM7.66055 17.0007V14.6247C7.66055 14.1823 7.47952 14.0102 7.01399 14.0102H5.95365C4.21227 14.0102 3.3933 13.1664 3.3933 11.5687V5.3011C3.3933 3.70346 4.21227 2.86778 5.95365 2.86778H17.0571C18.7899 2.86778 19.6175 3.70346 19.6175 5.3011V11.5687C19.6175 13.1664 18.7899 14.0102 17.0571 14.0102H11.4709C10.9881 14.0102 10.7467 14.0758 10.4192 14.3953L7.66055 17.0007ZM6.0571 9.38937H9.6088C9.88466 9.38937 10.1088 9.17641 10.1088 8.90598C10.1088 8.65206 9.88466 8.43903 9.6088 8.43903H6.0571C5.78123 8.43903 5.5571 8.65206 5.5571 8.90598C5.5571 9.17641 5.78123 9.38937 6.0571 9.38937ZM11.3933 9.38937H16.9709C17.2468 9.38937 17.4623 9.17641 17.4623 8.90598C17.4623 8.65206 17.2468 8.43903 16.9709 8.43903H11.3933C11.1174 8.43903 10.8933 8.65206 10.8933 8.90598C10.8933 9.17641 11.1174 9.38937 11.3933 9.38937ZM6.0571 11.5277H7.52262C7.79848 11.5277 8.014 11.323 8.014 11.0608C8.014 10.7904 7.79848 10.5774 7.52262 10.5774H6.0571C5.78123 10.5774 5.5571 10.7904 5.5571 11.0608C5.5571 11.323 5.78123 11.5277 6.0571 11.5277ZM9.29848 11.5277H13.6175C13.8933 11.5277 14.1174 11.323 14.1174 11.0608C14.1174 10.7904 13.8933 10.5774 13.6175 10.5774H9.29848C9.02262 10.5774 8.8071 10.7904 8.8071 11.0608C8.8071 11.323 9.02262 11.5277 9.29848 11.5277ZM15.4019 11.5277H16.9709C17.2468 11.5277 17.4623 11.323 17.4623 11.0608C17.4623 10.7904 17.2468 10.5774 16.9709 10.5774H15.4019C15.126 10.5774 14.9019 10.7904 14.9019 11.0608C14.9019 11.323 15.126 11.5277 15.4019 11.5277Z"
-                                fill="#A7A9B4"
-                              />
-                            </g>
-                            <defs>
-                              <clipPath id="clip0_1_6044">
-                                <rect
-                                  width="19"
-                                  height="18"
-                                  fill="white"
-                                  transform="translate(2.00537 0.5)"
-                                />
-                              </clipPath>
-                            </defs>
-                          </svg>
+                          <CommentIcon />
 
                           <span>{post.commentCount}</span>
                         </div>
