+++ src/components/post/comment/CommentItem.tsx
@@ -1,11 +1,11 @@
 import { ChangeEvent } from 'react';
 
+import { Button } from '@components/common/Button';
+import TextArea from '@components/common/TextArea';
+
 import { useAuthStore } from '@stores/useAuthStore';
 
 import { formatDateToYYMMDD } from '@utils/formatDateToYYMMDD';
-
-import { Button } from '../common/Button';
-import TextArea from '../common/TextArea';
 
 interface CommentItemProps {
   author: string;
